+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 0.2.0 (UNRELEASED)
 ------------------
 
+- Less false positives in IP address detection.
 - Officially support Python 3.4 too.
 - More strict checks for URI_IDs.